requirements.txt
pandas
ecg_plot
tqdm
numpy
torch
fastai
scipy
ipywidgets
pyyaml
PyWavelets
wfdb
hydra-core
omegaconf 